Basis for a websocket server and client.
NEW: A server that is intended to provide all the data necessary for the UI. Currently, it provides an interface through which leaf_percent, stem_percent, root_percent and starch_percent can be retrieved.

NEW: A client that serves as an interface between the UI and the server. At the same time, it can also serve as an interface for software that is designed to play the game without the UI. Currently, leaf_percent, stem_percent, root_percent and starch_percent can be retrieved using the client.