Client-server architecture (#3)
NEW: A server that is supposed to provide all the data needed for the user interface. It operates in its own thread and provides an interface through which leaf_percent, stem_percent, root_percent and starch_percent can be retrieved. In addition, the method 'calc_send_growth_rate' exists. This can calculate a new growth rate and send it back to all clients. The server can process multiple requests in one JSON, and the individual tasks no longer need to be submitted individually. Individual tasks are also possible.
NEW: A client that serves as an interface between the user interface and the server. At the same time, it can also serve as an interface for software that is to play the game without the user interface. Currently, the client can be used to query the values leaf_percent, stem_percent, root_percent and starch_percent. As well as the opening and closing of the stomata and the activation or deactivation of the starch pool as well as the query of information regarding the water supply. Finally, the method 'growth_rate' provides the results of a recalculation of the growth rates.
NEW: Use tox for the tests. Therefore the creation of a tox configuration file (tox.ini)
NEW: requirements.txt added
NEW: GrowthPercent class, which serves as a data container for all growth percentages.
NEW: GrowthRate class, which serves as a data container for all growth rates.
NEW: Water class. This contains all the information necessary for the UI regarding the water in the model.

UPDATED: game.py can now accept a command line argument that determines the log level.
UPDATED: ui.py no longer uses the dynamic model directly, but only through the client.
UPDATED (DynamicModel): This class now uses GrowthPercent instead of an array.
UPDATED (plant.py): This class now uses GrowthPercent.
UPDATED: dynamic_mode was given more log messages.
UPDATED: Removal of unnecessary pass-through of DynamicModel objects.
UPDATED: Removal of redundant files.
UPDATED: Restructuring as preparation for packet creation for PyPI.
UPDATED: Source Code Reformatted according to Black