From Growth to Plant
UPDATED: A plant is now sent instead of a growth object in response to a simulation step.

UPDATED: Some log sets have been adjusted. Furthermore, three log files are now created, one for the UI, one for the client and one for the server.

UPDATED: ServerPlant now processes with micromol instead of mol.