Variable port for the server and new test for water
NEW: A new exception indicates that negative biomass was used, although this should not be possible and thus indicates an error.

NEW: There is now a test for the water object.

UPDATED: The Water object now contains the methods to_dict and from_dict to create a dict analogous to the methods to_json and from_json. Furthermore, the use of negative biomass causes a NegativeBiomassError to be raised.

UPDATED: The Plant object no longer converts the contained Water object to a string for its own serialization to a JSON string but converts.

UPDATED: The server now uses a port dynamically and is no longer explicitly set to port 4000. Furthermore, the server can now accept connections from non-localhost. But this has to be activated explicitly by parameters. Moreover, the server has two additional variables, one that contains the port used and one that contains the ip_address.

UPDATED: The tests for the server object can now run in parallel, since the port is assigned dynamically and not fixed to 4000.

UPDATED: Version increased to 0.2.0.

BUGFIX: In game.py, the time of the last requested simulation is set. This was forgotten, so the time difference was equal to the runtime.