Adaptation so that the server process also runs under Windows
NEW: there is a container for the server that can be used from outside so that the server works directly in its own process and thus runs independently of the main process. This container takes care of providing some information about the server, such as its port and IP address.

UPDATED: The GitHub action for the tests now runs on every pull request and without the need for syntax checks.

UPDATED; The tests for the server have been adjusted so that they work with the other changes.