Nitrate from the soil
UPDATED: The nitrate grid of the environment is now also included in the calculation of the available nitrate for the FBA.

NEW: The client receives the method 'create_new_first_letter', this can be used to use the method of the same name of the root/LSystems via the client/server and in this way have an influence on the server-side root.

NEW: The client receives the function 'add2grid'. This can be used to add water or nitrates to a cell in the representation of the soil.

Fix: Correct casting of Float to Int for the hour calculation which is necessary for the weather simulation.