Refactoring and new tests.
NEW: Two new tests. They call a growth command every second using the client, once for 30 seconds and once for 30 minutes. They serve as a stress test.

UPDATED: Added time.sleep to allow the client thread to work.

UPDATED: The client will now kill itself when the server closes the connection. In addition, it is able to handle it when multiple requests overlap.

UPDATED: Changes after Flake8 and Black.