Split GitHub Actions
UPDATED: The GitHub Action workflow files have been split into three parts. The first workflow consists of flake8, black and Mypy. The second consists of the tests and the third creates the package, creates a tag if necessary and uploads it to PyPI. The first two workflows are always executed on pull requests. The third one is only executed for changes in the main branch.