Environment & Weather (#12)
### Environment
NEW: Environment class. This includes the nitrates and water occurrence in the soil.

NEW: MetaboliteGrid is a class that describes a metabolite present in the soil and defines its behavior and availability to the plant.


UPDATED: Environment now uses WeatherStates to describe the hourly states.

### Weather
NEW: WeatherSimulatorMinimal this class contains all WeatherStates and the info up to which hour has been simulated so far.

NEW: WeatherStates class. This describes the weather state for one hour.

### Server
NEW: The server has the option to reload the environment and reset the state of the installation.

NEW: ServerPlant now also contains an LSystem object representing the root structure.

NEW: The Server now contains an environment object that defines the water and nitrate occurrence in the ground.


UPDATED: The server now contains an environment object and the possibility to query it.

BUGFIX: The server handler can now successfully deserialize the leaf creation request.

### Client
NEW: Client has the method load_level. This can be used to reset the server's plant object and reload the environment.

NEW: The client receives the method 'create_new_first_letter'; this can be used to use the method of the same name of the root/LSystems via the client/server and, in this way, influence the server-side root.

NEW: The client receives the function 'add2grid'. This can be used to add water or nitrates to a cell in the representation of the soil.

UPDATED: The client reports not receiving tasks.

UPDATED: The client can request the environment object of the server via 'get_environment'.

UPDATED: The client will now stop when the server closes the connection. In addition, it is able to handle it when multiple requests overlap.

### FBA/DynamicModel
UPDATED: DynamicModel no longer contains the transpiration factor calculation. Furthermore, the boost has been removed, and the biomass conversion has been removed, as the model directly outputs grams for the biomasses.

UPDATED: The nitrate grid of the environment is now also included in the calculation of the available nitrate for the FBA.

UPDATED: dynamic_model now uses the specification of VMax from constants. Also, the photon upper limit is now scaled correctly. Last but not least, biomass production is now scaled according to the sum of the available space of each leaf object. This means that the model can only create as much leaf biomass as it can fit into the previously created leaves.

### Water
UPDATED: water.py has a method that specifies how much micromol can be added to the pool.

UPDATED: water now also contains the calculation of the transpiration factor.

BUGFIX: The initial amount of water can now be reduced without reducing to max value. Therefore, for example, you can start with 100 micromoles as before but reduce it to 90 even if 50 is the maximum. Previously, the pool would have been automatically reduced to 50 at the next change.

### Plant/Leaf
NEW: 'Leaf' class, which contains all server-side relevant information about a leaf, such as a unique ID, current weight, and maximum weight. This class is used by plants and leaves.

NEW: 'Leafs' This class represents all individual leaves and takes care of providing all information related to the leaves. Exact information on the individual leaves can be found in the set of individual leaves.

UPDATED: The Plant class has passed all leaf-related responsibilities to the leafs class. Accordingly, all methods associated with leaves are now in this class and no longer directly in Plant.

### Tests

NEW: Two new tests. They call a growth command every second using the client, once for 30 seconds and once for 30 minutes. They serve as a stress test.

UPDATED: Added tests for the client, with a new test case for creating a leaf. Added mock for leaf ID so that it is clear which ID the first leaf automatically created by the init of plant.py has.

UPDATED: All existing test classes have been revised. This includes additional cases as well as adaptations to changes in the code. 'test_slider' was temporarily removed because it caused errors due to tox's isolated test environment and pygames' dependency on graphics and audio devices.

### Other

UPDATED: The plant class takes into account that the FBA model outputs Gram directly.

UPDATED: LSystem now only imports pygame locally to avoid problems when importing pygame on machines without sound or graphic output.

UPDATED: Biomass factor removed from constants.py because the model outputs grams directly.

UPDATED: game.py now contains a copy of the server environment.

UPDATED: Added time.sleep to allow the client thread to work.

BUGFIX: Nitrate usage can no longer exceed the available amount in the soil.

BUGFIX: Changes to nitrates reversed, as these nitrates potentially dropped consumption below the table. Also fixed a bug that caused nitrates to be subtracted twice from the internal pool, resulting in negative values.

BUGFIX: Roots are recreated correctly.

UPDATED: LSystems is now comparable.

UPDATED: Root calc_position is automatically executed as soon as the biomass of the plant object changes. Therefore, removed from dynamic_model.

UPDATED: The GitHub Action workflow files have been split into three parts. The first workflow consists of flake8, black, and Mypy. The second consists of the tests, and the third creates the package, creates a tag if necessary, and uploads it to PyPI. The first two workflows are always executed on pull requests. The third one is only executed for changes in the main branch.

UPDATED: Several methods were removed because they were not used.

UPDATED: Adjustments in accordance with black, flake8, and Mypy.

UPDATED: Badges added to the README

BUGFIX: Corrected unit calculation of photons. Due to a unit error in the SLA, multiplication and division are now obsolete.

BUGFIX: The strength percentage was still divided by 100, even though the percentages stored in GrowthPercent objects are already a decimal number between 0 and 1 and, therefore, do not need to be adjusted.