Fix for model constraints and UI percentages. (#13)
* BUGFIX: Constraints & Percentages

BUGFIX: The GitHub action 'publish' no longer requires non-existing tasks as a prerequisite. Also, temporarily changed to an echo instead of a real upload.

BUGFIX: dynamic_model adjusts the constraints in each iteration. Previously, this was only done for changes in the growth percentage ratios. However, since the mass of the individual organs plays a role within the constraints and generally changes unevenly in each iteration, the adjustment is thus usually always necessary.

BUGFIX: In 'game.py', the 'GrowthPercent' object is now created using the decimal notation of the percentages. This happened before against the object description with values between 1 and 100. Furthermore, the windowed flag works again.

NEW: The client gets a new function called flush. This allows to force the execution of scheduled tasks. This can use a timeout to allow the client to execute for a limited time or not, which will block the executing thread until all tasks are finished.

NEW: Test for the flush method of the client.

UPDATED: The game loop in game.py now uses the new flush method of the client.

UPDATED: The growth_rate request is no longer transmitted as a string but is temporarily stored as dict before being serialized to JSON. The deserialization was adapted analogously.

UPDATED: The 'pyproject.toml' was changed to Python version 3.11. As well as the version of PlantEd was increased to 0.4.0. The requirements.txt has been updated as well for Python 3.11.

UPDATED: The name of the development environment within Tox has been changed from 'py310-req' to 'py-dev'.

UPDATED: The README was extended to include start options for PlantEd and an explanation of the development environments.

UPDATED: The SLA value was moved from the Leaf object to the dynamic_model object.

REMOVED: Test for calculating the specific leaf area, as this method has been removed.